+++ web/app/themes/theme/archive.php
@@ -1,5 +1,5 @@
 <?php get_template_part('templates/page', 'header'); ?>
-
+<h1>Archive Template</h1>
 <?php if (!have_posts()) : ?>
   <div class="alert alert-warning">
     <?php _e('Sorry, no results were found.', 'sage'); ?>