+++ web/app/themes/theme/lib/post-types.php
@@ -5,13 +5,13 @@
   Custom post types and taxonomies
   --------------------------------
 
-  Post Types
+  1. Post Types
   ==========
 
   Event:           'post_type' => 'event'
   Story:           'post_type' => 'story'
 
-  Taxonomies
+  2. Taxonomies
   ==========
 
   Event Timeline:   event_timeline
@@ -31,23 +31,31 @@ add_action( 'init', 'register_post_types' );
 add_action( 'init', 'register_taxonomies' );
 
 
-// Post types
-// ----------
+// 1. Post types
+// --------------
 
 function register_post_types() {
 
-  // Eve
+  // Event
   // =====
 
   $post_type      = 'event';
-  $slug           = 'history/events';
-  $supported_ui   = array( 'title', 'thumbnail' ); // false to hide all default WordPress post editing UI
 
+  $slug           = get_field('event_post_type_slug_base', 'option') . get_field('event_post_type_slug', 'option');
+  $supported_ui   = ['title', 'thumbnail']; // false to hide all default WordPress post editing UI
+
+  // **Note on event slug**
+
+  // Event post slugs and slug base should be on Site settings page
+  // Event **slug base** e.g. "history/" is used for better semantics of the event URL
+  // E.g. naming a slug base 'history' and event slug 'events' will result in 'history/events'.
+  // Event **slug base** is also used as a slug for event post type taxonomies. E.g. 'history/taxonomy/taxomomy-term'.
+
+  // Labels
   $singular       = 'Event';
   $plural         = 'Events';
 
-
-  $labels = array(
+  $labels = [
     'name'                        => __( $plural ),
     'singular_name'               => __( $singular ),
     'menu_name'                   => __( $plural ),
@@ -62,35 +70,36 @@ function register_post_types() {
     'parent_item_colon'           => __( 'Parent ' . $plural . ':' ),
     'not_found'                   => __( 'No ' . $plural . ' found.' ),
     'not_found_in_trash'          => __( 'No ' . $plural . ' found in Trash.' )
-  );
+  ];
 
-  $args = array(
-    'hierarchical'                => true,
+  $args = [
     'has_archive'                 => false, // set 'true' to use archive-post-type.php template
-    'capability_type'             => 'post',
-    'rewrite'                     => array('slug' => $slug, 'with_front' => false),
+    'public'                      => true,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
     'supports'                    => $supported_ui,
     'labels'                      => $labels,
+    'capability_type'             => 'post',
     'menu_position'               => 6, // menu order overwritten in admin-cp.php
     'menu_icon'                   => 'dashicons-clock',
     'show_ui'                     => true,
-    'query_var'                   => true,
-    'public'                      => true
-  );
+  ];
 
   register_post_type( $post_type, $args );
+
 
   // Story
   // =====
 
   $post_type      = 'story';
-  $slug           = 'stories';
-  $supported_ui   = array( 'post-formats', 'thumbnail' ); // false to hide all default WordPress post editing UI
 
+  $slug           = get_field('story_slug', 'option');
+  $supported_ui   = ['title', 'thumbnail']; // false to hide all default WordPress post editing UI
+
+  // Labels
   $singular       = 'Story';
   $plural         = 'Stories';
 
-  $labels = array(
+  $labels = [
     'name'                        => __( $plural ),
     'singular_name'               => __( $singular ),
     'menu_name'                   => __( $plural ),
@@ -105,43 +114,63 @@ function register_post_types() {
     'parent_item_colon'           => __( 'Parent ' . $plural . ':' ),
     'not_found'                   => __( 'No ' . $plural . ' found.' ),
     'not_found_in_trash'          => __( 'No ' . $plural . ' found in Trash.' )
-  );
+  ];
 
-  $args = array(
-    'hierarchical'                => true,
-    'has_archive'                 => false, // set 'true' to use archive-post-type.php template
-    'capability_type'             => 'post',
-    'rewrite'                     => array('slug' => $slug, 'with_front' => false),
+  $args = [
+    //'has_archive'                 => false, // set 'true' to use archive-post-type.php template
+    'public'                      => true,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => false],
     'supports'                    => $supported_ui,
     'labels'                      => $labels,
+    'capability_type'             => 'post',
     'menu_position'               => 6, // menu order overwritten in admin-cp.php
-    'menu_icon'                   => 'dashicons-format-status',
+    'menu_icon'                   => 'dashicons-clock',
     'show_ui'                     => true,
-    'query_var'                   => true,
-    'public'                      => true
-  );
+  ];
 
   register_post_type( $post_type, $args );
+
 }
 
-// Taxonomies
-// ----------
+/*
+  Taxonomies
+  ----------
+
+  Note on taxonomies slugs:
+
+  All event taxonomies use the same slug which is used instead of individual taxonomy slug
+  For example: defining 'events' slug (note plural case) for all event post taxonomies will result in:
+
+ ../events/united-states            (for 'timeline' taxonomy term)
+ ../events/identity-and-belonging   (for 'topic' taxonomy term)
+
+
+ Defining 'stories' slug for all story post taxonomies will result in:
+
+  ../stories/new-york               (for 'city' taxonomy term)
+ ../stories/growing-up              (for 'topic' taxonomy term)
+
+ Note: avoid having identical post type and taxonomy slugs. Use singular slugs for custom post types
+ and plural for taxonomies. Eg. 'event' for custom post type and 'events' slug for this post type taxonomies
+
+ Having identical slugs for post types and their taxonomies will result in 'page not found' error.
+
+*/
+
 
 function register_taxonomies() {
 
   // Event Timeline
   // ==============
 
-  // Note: this taxonomy is for internal use only and it is not publicly queryable.
-
   $taxonomy       = 'event_timeline';
-  //$slug         = 'history/timeline';
+  $slug           = get_field('event_post_type_slug_base', 'option') . get_field('one_slug_for_all_event_post_type_taxonomies', 'option');
   $post_types     = 'event';
 
   $plural         = 'Timelines';
   $singular       = 'Timeline';
 
-  $labels = array(
+  $labels = [
     'name'                        => __( $plural ),
     'singular_name'               => __( $singular ),
     'search_items'                => __( 'Search ' . $plural ),
@@ -153,18 +182,21 @@ function register_taxonomies() {
     'add_new_item'                => __( 'Add New ' . $singular ),
     'new_item_name'               => __( 'New ' . $singular ),
     'menu_name'                   => __( $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'                => true,
-    'public'                      => false,
+  $args = [
     'labels'                      => $labels,
+
+    'public'                      => false,
     'show_ui'                     => true,
-    'meta_box_cb'                 => false, // true to show meta box on post edit page
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
     'show_admin_column'           => true,
-    'query_var'                   => true,
-    //'rewrite'                   => array('slug' => $slug, 'with_front' => false)
-  );
+
+    'hierarchical'                => true,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true, 'hierarchical'=> true],
+    'sort'                        => true, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -175,13 +207,13 @@ function register_taxonomies() {
   // Note: this taxonomy is for internal use only and it is not publicly queryable.
 
   $taxonomy       = 'event_era';
-  //$slug         = 'history/events/eras';
+  $slug           = get_field('event_post_type_slug_base', 'option') . get_field('one_slug_for_all_event_post_type_taxonomies', 'option');
   $post_types     = 'event';
 
   $singular       = 'Era';
   $plural         = 'Eras';
 
-  $labels = array(
+  $labels = [
     'name'                        => __( $plural ),
     'singular_name'               => __( $singular ),
     'search_items'                => __( 'Search ' . $plural ),
@@ -193,18 +225,21 @@ function register_taxonomies() {
     'add_new_item'                => __( 'Add New ' . $singular ),
     'new_item_name'               => __( 'New ' . $singular ),
     'menu_name'                   => __( $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'                => true,
-    'public'                      => false,
+  $args = [
     'labels'                      => $labels,
+
+    'public'                      => true,
     'show_ui'                     => true,
-    'meta_box_cb'                 => false, // true to show meta box on post edit page
-    'show_admin_column'           => false,
-    'query_var'                   => true,
-    //'rewrite'                   => array('slug' => $slug, 'with_front' => false)
-  );
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => false, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -213,13 +248,13 @@ function register_taxonomies() {
   // ==========
 
   $taxonomy       = 'event_type';
-  $slug           = 'history/events/types';
+  $slug           = get_field('event_post_type_slug_base', 'option') . get_field('one_slug_for_all_event_post_type_taxonomies', 'option');
   $post_types     = 'event';
 
   $singular       = 'Type';
   $plural         = 'Types';
 
-  $labels = array(
+  $labels = [
     'name'                       => __( $plural ),
     'singular_name'              => __( $singular ),
     'search_items'               => __( 'Search ' . $plural ),
@@ -236,17 +271,21 @@ function register_taxonomies() {
     'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
     'not_found'                  => __( 'No ' . $plural  .' found.' ),
     'menu_name'                  => __(  $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'               => false,
-    'labels'                     => $labels,
-    'show_ui'                    => true,
-    'meta_box_cb'                => false, // true to show meta box on post edit page
-    'show_admin_column'          => false,
-    'query_var'                  => true,
-    'rewrite'                    => array('slug' => $slug, 'with_front' => false)
-  );
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => false, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -255,13 +294,13 @@ function register_taxonomies() {
   // ===========
 
   $taxonomy       = 'event_topic';
-  $slug           = 'history/events/topics';
+  $slug           = get_field('event_post_type_slug_base', 'option') . get_field('one_slug_for_all_event_post_type_taxonomies', 'option');
   $post_types     = 'event';
 
   $singular       = 'Topic';
   $plural         = 'Topics';
 
-  $labels = array(
+  $labels = [
     'name'                       => __( $plural ),
     'singular_name'              => __( $singular ),
     'search_items'               => __( 'Search ' . $plural ),
@@ -278,18 +317,21 @@ function register_taxonomies() {
     'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
     'not_found'                  => __( 'No ' . $plural  .' found.' ),
     'menu_name'                  => __(  $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'               => false,
-    'sort'                       => true,
-    'labels'                     => $labels,
-    'show_ui'                    => true,
-    'meta_box_cb'                => false, // true to show meta box on post edit page
-    'show_admin_column'          => false,
-    'query_var'                  => true,
-    'rewrite'                    => array('slug' => $slug, 'with_front' => false)
-  );
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => true,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => false],
+    'sort'                        => true, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -298,13 +340,13 @@ function register_taxonomies() {
   // ===========
 
   $taxonomy       = 'event_group';
-  $slug           = 'history/events/groups';
+  $slug           = get_field('event_post_type_slug_base', 'option') . get_field('one_slug_for_all_event_post_type_taxonomies', 'option');
   $post_types     = 'event';
 
   $singular       = 'Group';
   $plural         = 'Groups';
 
-  $labels = array(
+  $labels = [
     'name'                       => __( $plural ),
     'singular_name'              => __( $singular ),
     'search_items'               => __( 'Search ' . $plural ),
@@ -321,18 +363,21 @@ function register_taxonomies() {
     'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
     'not_found'                  => __( 'No ' . $plural  .' found.' ),
     'menu_name'                  => __(  $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'               => false,
-    'sort'                       => true,
-    'labels'                     => $labels,
-    'show_ui'                    => true,
-    'meta_box_cb'                => false, // true to show meta box on post edit page
-    'show_admin_column'          => false,
-    'query_var'                  => true,
-    'rewrite'                    => array('slug' => $slug, 'with_front' => false)
-  );
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => true, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -341,13 +386,13 @@ function register_taxonomies() {
   // ===========
 
   $taxonomy       = 'story_topic';
-  $slug           = 'stories/topics';
+  $slug           = get_field('one_slug_for_all_story_post_type_taxonomies', 'option');
   $post_types     = 'story';
 
   $singular       = 'Topic';
   $plural         = 'Topics';
 
-  $labels = array(
+  $labels = [
     'name'                       => __( $plural ),
     'singular_name'              => __( $singular ),
     'search_items'               => __( 'Search ' . $plural ),
@@ -364,18 +409,67 @@ function register_taxonomies() {
     'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
     'not_found'                  => __( 'No ' . $plural  .' found.' ),
     'menu_name'                  => __(  $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'               => false,
-    'sort'                       => true,
-    'labels'                     => $labels,
-    'show_ui'                    => true,
-    'meta_box_cb'                => false, // true to show meta box on post edit page    'meta_box_cb'       => false, // true to show meta box on post edit page
-    'show_admin_column'          => false,
-    'query_var'                  => true,
-    'rewrite'                    => array('slug' => $slug, 'with_front' => false)
-  );
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => false, // whether taxonomy should remember the order in which terms are added to objects
+  ];
+
+  register_taxonomy( $taxonomy, $post_types, $args );
+
+
+  // Story City
+  // ===========
+
+  $taxonomy       = 'story_city';
+  $slug           = get_field('one_slug_for_all_story_post_type_taxonomies', 'option');
+  $post_types     = 'story';
+
+  $singular       = 'City';
+  $plural         = 'Cities';
+
+  $labels = [
+    'name'                       => __( $plural ),
+    'singular_name'              => __( $singular ),
+    'search_items'               => __( 'Search ' . $plural ),
+    'popular_items'              => __( 'Popular ' . $plural ),
+    'all_items'                  => __( 'All ' . $plural ),
+    'parent_item'                => null,
+    'parent_item_colon'          => null,
+    'edit_item'                  => __( 'Edit ' . $singular  ),
+    'update_item'                => __( 'Update ' . $singular  ),
+    'add_new_item'               => __( 'Add New ' . $singular ),
+    'new_item_name'              => __( 'Add New ' . $singular . ' Name'),
+    'separate_items_with_commas' => __( 'Separate ' . $plural . ' with commas' ),
+    'add_or_remove_items'        => __( 'Add or remove ' . $plural . ' writers' ),
+    'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
+    'not_found'                  => __( 'No ' . $plural  .' found.' ),
+    'menu_name'                  => __(  $plural ),
+  ];
+
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => false, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -383,16 +477,14 @@ function register_taxonomies() {
   // Story Group
   // ===========
 
-  // Note: this taxonomy is for internal use only and it is not publicly queryable.
-
   $taxonomy       = 'story_group';
-  //$slug         = 'stories/group';
+  $slug           = get_field('one_slug_for_all_story_post_type_taxonomies', 'option');
   $post_types     = 'story';
 
   $singular       = 'Group';
   $plural         = 'Groups';
 
-  $labels = array(
+  $labels = [
     'name'                       => __( $plural ),
     'singular_name'              => __( $singular ),
     'search_items'               => __( 'Search ' . $plural ),
@@ -409,18 +501,21 @@ function register_taxonomies() {
     'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
     'not_found'                  => __( 'No ' . $plural  .' found.' ),
     'menu_name'                  => __(  $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'               => false,
-    'public'                     => false,
-    'labels'                     => $labels,
-    'show_ui'                    => true,
-    'meta_box_cb'                => false, // true to show meta box on post edit page    'meta_box_cb'       => false, // true to show meta box on post edit page
-    'show_admin_column'          => true,
-    'query_var'                  => true,
-    //'rewrite'                  => array('slug' => $slug, 'with_front' => false)
-  );
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => false, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 
@@ -431,13 +526,13 @@ function register_taxonomies() {
   // Note: this taxonomy is shared beetween event and story post types
 
   $taxonomy       = 'global_tag';
-  $slug           = 'tags';
-  $post_types     = false; // array( 'event', 'story' );
+  $slug           = get_field('global_tag_slug', 'option');
+  $post_types     = ['event', 'story'];
 
   $singular       = 'Tag';
   $plural         = 'Tags';
 
-  $labels = array(
+  $labels = [
     'name'                       => __( $plural ),
     'singular_name'              => __( $singular ),
     'search_items'               => __( 'Search ' . $plural ),
@@ -454,18 +549,22 @@ function register_taxonomies() {
     'choose_from_most_used'      => __( 'Choose from the most used ' . $plural ),
     'not_found'                  => __( 'No ' . $plural  .' found.' ),
     'menu_name'                  => __(  $plural ),
-  );
+  ];
 
-  $args = array(
-    'hierarchical'               => false,
-    'sort'                       => true,
-    'labels'                     => $labels,
-    'show_ui'                    => true,
-    'show_admin_column'          => true,
-    'update_count_callback'      => '_update_post_term_count',
-    'query_var'                  => true,
-    'rewrite'                    => array( 'slug' => $slug ),
-  );
+  $args = [
+    'labels'                      => $labels,
+
+    'public'                      => true,
+    'show_ui'                     => true,
+    'show_in_nav_menus'           => false,
+    'meta_box_cb'                 => false, // hide show meta box on post edit page
+    'show_admin_column'           => true,
+
+    'hierarchical'                => false,
+    'update_count_callback'       => '_update_post_term_count',
+    'rewrite'                     => ['slug' => $slug, 'with_front' => true],
+    'sort'                        => true, // whether taxonomy should remember the order in which terms are added to objects
+  ];
 
   register_taxonomy( $taxonomy, $post_types, $args );
 }