+++ web/app/themes/theme/single-story.php
@@ -1,0 +1,1 @@
+<?php get_template_part('templates/content-story', get_post_type()); ?>