+++ web/app/themes/theme/template-home.php
@@ -1,7 +1,0 @@
-<?php
-/**
- * Template Name: Homepage
- */
-?>
-
-<?php print 'This is homepage template' ?>